SAC-28992: Refactored existing streams code
