+++ tap_zendesk/streams/groups.py
@@ -1,0 +1,24 @@
+from singer import utils
+from tap_zendesk.streams.abstracts import (
+    CursorBasedStream
+)
+
+
+class Groups(CursorBasedStream):
+    name = "groups"
+    replication_method = "INCREMENTAL"
+    replication_key = "updated_at"
+    endpoint = 'https://{}.zendesk.com/api/v2/groups'
+    item_key = 'groups'
+
+    def sync(self, state):
+        bookmark = self.get_bookmark(state)
+
+        groups = self.get_objects()
+        for group in groups:
+            if utils.strptime_with_tz(group['updated_at']) >= bookmark:
+                # NB: We don't trust that the records come back ordered by
+                # updated_at (we've observed out-of-order records),
+                # so we can't save state until we've seen all records
+                self.update_bookmark(state, group['updated_at'])
+                yield (self.stream, group)
