+++ tap_zendesk/streams/organizations.py
@@ -1,0 +1,47 @@
+import datetime
+from zenpy.lib.exception import APIException
+from tap_zendesk.streams.abstracts import (
+    Stream,
+    process_custom_field,
+    raise_or_log_zenpy_apiexception,
+    START_DATE_FORMAT
+)
+
+
+class Organizations(Stream):
+    name = "organizations"
+    replication_method = "INCREMENTAL"
+    replication_key = "updated_at"
+    endpoint = 'https://{}.zendesk.com/api/v2/organizations'
+    item_key = 'organizations'
+
+    def _add_custom_fields(self, schema):
+        endpoint = self.client.organizations.endpoint
+        # NB: Zenpy doesn't have a public endpoint for this at time of writing
+        #     Calling into underlying query method to grab all fields
+        try:
+            field_gen = self.client.organizations._query_zendesk(endpoint.organization_fields, # pylint: disable=protected-access
+                                                                 'organization_field')
+        except APIException as e:
+            return raise_or_log_zenpy_apiexception(schema, self.name, e)
+        schema['properties']['organization_fields']['properties'] = {}
+        for field in field_gen:
+            schema['properties']['organization_fields']['properties'][field.key] = process_custom_field(field)
+
+        return schema
+
+    def sync(self, state):
+        bookmark = self.get_bookmark(state)
+        organizations = self.client.organizations.incremental(start_time=bookmark)
+        for organization in organizations:
+            self.update_bookmark(state, organization.updated_at)
+            yield (self.stream, organization)
+
+    def check_access(self):
+        '''
+        Check whether the permission was given to access stream resources or not.
+        '''
+        # Convert datetime object to standard format with timezone. Used utcnow to reduce API call burden at discovery time.
+        # Because API will return records from now which will be very less
+        start_time = datetime.datetime.utcnow().strftime(START_DATE_FORMAT)
+        self.client.organizations.incremental(start_time=start_time)
