+++ tap_zendesk/streams/sla_policies.py
@@ -1,0 +1,16 @@
+from tap_zendesk.streams.abstracts import Stream
+
+
+class SLAPolicies(Stream):
+    name = "sla_policies"
+    replication_method = "FULL_TABLE"
+
+    def sync(self, state): # pylint: disable=unused-argument
+        for policy in self.client.sla_policies():
+            yield (self.stream, policy)
+
+    def check_access(self):
+        '''
+        Check whether the permission was given to access stream resources or not.
+        '''
+        self.client.sla_policies()
