+++ tap_zendesk/streams/ticket_fields.py
@@ -1,0 +1,24 @@
+from singer import utils
+from tap_zendesk.streams.abstracts import (
+    CursorBasedStream
+)
+
+
+class TicketFields(CursorBasedStream):
+    name = "ticket_fields"
+    replication_method = "INCREMENTAL"
+    replication_key = "updated_at"
+    endpoint = 'https://{}.zendesk.com/api/v2/ticket_fields'
+    item_key = 'ticket_fields'
+
+    def sync(self, state):
+        bookmark = self.get_bookmark(state)
+
+        fields = self.get_objects()
+        for field in fields:
+            if utils.strptime_with_tz(field['updated_at']) >= bookmark:
+                # NB: We don't trust that the records come back ordered by
+                # updated_at (we've observed out-of-order records),
+                # so we can't save state until we've seen all records
+                self.update_bookmark(state, field['updated_at'])
+                yield (self.stream, field)
